Directly load completed issues.
You can link to completed issues, but these are not always already loaded, so they may not be in @issueCollection. Therefore, if this is the case, try to download the issue directly through the API. When doing so, also show a 'loading..' overlay on the issue view to inform the user we are trying to do something.